HELP!! Added insert_into_viewed_by to UserCtrl, but have cursor
not connected-error. Cannot find the problem right now, but you might
spot it quickly ;)) I think we should reconsider how often and where we
open and close connections and cursors in the future. Strange that it
fails that far up in the main-file, since it has always worked there.
Weird...