Started on uc5 (#6)
* started on uc5

* Made some minor changes + added some test-data.

The query looks to work just fine!

* Added an ORDER BY readPosts, writtenPosts (in case of tie).

* HELP!! Added insert_into_viewed_by to UserCtrl, but have cursor

not connected-error. Cannot find the problem right now, but you might
spot it quickly ;)) I think we should reconsider how often and where we
open and close connections and cursors in the future. Strange that it
fails that far up in the main-file, since it has always worked there.
Weird...

* Insert into ViewedBy works in the other UC's I think.

Please check yourselves as well ;)

* Added my solution for closing connections. Left print in dels

in order for you to see that the destructors are called at the correct
times. Please remove these if you agree that this is the best solution
to the problem!

* Removed print statements in destr.

Co-authored-by: alexaoh <alexaoh@users.noreply.github.com>